fix(ci): #841 CI 失敗を解消
